moved player out into its own component and passed player HP and playerPokemonObject down as props.
