+++ src/Foe/Foe.js
